+++ odoo/local-src/specific_sale/models/sale.py
@@ -127,6 +127,11 @@ class SaleOrderLine(models.Model):
     supplier_break = fields.Boolean(compute='_compute_supplier_break')
     exception = fields.Char(compute='_compute_exception')
     date_order = fields.Datetime(related="order_id.date_order")
+    older_lot_life_date = fields.Datetime(
+        string='Expiration date',
+        related='product_id.older_lot_id.life_date',
+        readonly=True
+    )
 
     @api.depends('product_id')
     def _compute_supplier_break(self):